backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
loguru==0.7.2
orjson==3.10.7
